// src/frontend/src/app/api/models/create-daily-activity-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActivityStatus } from '../models/activity-status';
export interface CreateDailyActivityRequest {
  content: string;
  isBlocked: boolean;
  partnerId: string;
  referenceDate: string;
  status: ActivityStatus;
  theme: string;
  title: string;
}